pyproject.toml
[project]
name = "ai-resume-analyser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dateparser>=1.2.2",
    "langdetect>=1.0.9",
    "llama-index>=0.12.52",
    "llama-index-llms-openrouter>=0.3.2",
    "llama-index-tools-tavily-research>=0.3.0",
    "loguru>=0.7.3",
    "pdfplumber>=0.11.7",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "pymupdf>=1.26.3",
    "pypdf>=5.8.0",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "spacy>=3.8.7",
]
